requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
plotly>=5.18.0
reportlab>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

